Enhance documentation and structure of registration code
Added detailed XML comments for interfaces and classes to improve code readability and maintainability. Defined `IAverageView` and `IView` interfaces with clear contracts for updating registration details. Updated `UpdateRegistrationAveragePresenter` and `UpdateRegistrationDivisionPresenter` classes with comprehensive method summaries. Modified `TournamentRegistrationViewModel` for better testability. Removed unused `using` directives and organized namespaces for improved structure.
