+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Registrations/Retrieve/TournamentRegistrationViewModel.cs
@@ -1,36 +1,56 @@
-﻿
-namespace BowlingMegabucks.TournamentManager.Registrations.Retrieve;
+﻿namespace BowlingMegabucks.TournamentManager.Registrations.Retrieve;
+
+/// <summary>
+/// Represents a view model for a tournament registration.
+/// </summary>
 internal class TournamentRegistrationViewModel 
     : ITournamentRegistrationViewModel
 {
+    /// <inheritdoc/>
     public RegistrationId Id { get; }
 
+    /// <inheritdoc/>
     public string FirstName { get; init; }
 
+    /// <inheritdoc/>
     public string LastName { get; init; }
 
+    /// <inheritdoc/>
     public BowlerId BowlerId { get; }
 
+    /// <inheritdoc/>
     public string BowlerName { get; set; }
 
+    /// <inheritdoc/>
     public DivisionId DivisionId { get; }
 
+    /// <inheritdoc/>
     public string DivisionName { get; }
 
+    /// <inheritdoc/>
     public IEnumerable<SquadId> SquadsEntered { get; }
 
+    /// <inheritdoc/>
     public short SquadsEnteredCount
         => (short)SquadsEntered.Count();
 
+    /// <inheritdoc/>
     public IEnumerable<SquadId> SweepersEntered { get; }
 
+    /// <inheritdoc/>
     public short SweepersEnteredCount
         => (short)SweepersEntered.Count();
 
+    /// <inheritdoc/>
     public bool SuperSweeperEntered { get; set; }
 
+    /// <inheritdoc/>
     public int? Average { get; set; }
 
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TournamentRegistrationViewModel"/> class from a registration model.
+    /// </summary>
+    /// <param name="registration">The registration model.</param>
     public TournamentRegistrationViewModel(Models.Registration registration)
     {
         Id = registration.Id;
@@ -47,7 +67,7 @@ internal class TournamentRegistrationViewModel
     }
 
     /// <summary>
-    /// Unit Test Constructor
+    /// Initializes a new instance of the <see cref="TournamentRegistrationViewModel"/> class for unit testing.
     /// </summary>
     internal TournamentRegistrationViewModel()
     {
@@ -60,31 +80,73 @@ internal class TournamentRegistrationViewModel
     }
 }
 
-internal interface ITournamentRegistrationViewModel
+/// <summary>
+/// Represents the interface for a tournament registration view model.
+/// </summary>
+public interface ITournamentRegistrationViewModel
 {
+    /// <summary>
+    /// Gets the bowler's first name.
+    /// </summary>
     string FirstName { get; }
 
+    /// <summary>
+    /// Gets the bowler's last name.
+    /// </summary>
     string LastName { get; }
 
+    /// <summary>
+    /// Gets the registration identifier.
+    /// </summary>
     RegistrationId Id { get; }
 
+    /// <summary>
+    /// Gets the bowler identifier.
+    /// </summary>
     BowlerId BowlerId { get; }
 
+    /// <summary>
+    /// Gets or sets the bowler's display name.
+    /// </summary>
     string BowlerName { get; set; }
 
+    /// <summary>
+    /// Gets the division identifier.
+    /// </summary>
     DivisionId DivisionId { get; }
 
+    /// <summary>
+    /// Gets the division name.
+    /// </summary>
     string DivisionName { get; }
 
+    /// <summary>
+    /// Gets the squads the bowler is entered in.
+    /// </summary>
     IEnumerable<SquadId> SquadsEntered { get; }
 
+    /// <summary>
+    /// Gets the count of squads entered.
+    /// </summary>
     short SquadsEnteredCount { get; }
 
+    /// <summary>
+    /// Gets the sweepers the bowler is entered in.
+    /// </summary>
     IEnumerable<SquadId> SweepersEntered { get; }
 
+    /// <summary>
+    /// Gets the count of sweepers entered.
+    /// </summary>
     short SweepersEnteredCount { get; }
 
+    /// <summary>
+    /// Gets or sets a value indicating whether the bowler is a super sweeper.
+    /// </summary>
     bool SuperSweeperEntered { get; set; }
 
+    /// <summary>
+    /// Gets or sets the average score.
+    /// </summary>
     int? Average { get; set; }
 }
